# backend/docs/swagger.yaml
definitions:
...
  domain.AnalysisConfluenceResp:
    properties:
      content:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
...
  domain.EpubResp:
    properties:
      content:
        type: string
      title:
        type: string
    type: object
...
  domain.GetDocsReq:
    properties:
      integration:
        type: string
      kb_id:
        type: string
      pages:
        items:
          $ref: '#/definitions/domain.PageInfo'
        type: array
    required:
    - kb_id
    type: object
  domain.GetDocxReq:
    properties:
      app_id:
        type: string
      app_secret:
        type: string
      kb_id:
        type: string
      sources:
        items:
          $ref: '#/definitions/domain.Source'
        type: array
      user_access_token:
        type: string
    required:
    - kb_id
    - sources
    type: object
  domain.GetDocxResp:
    properties:
      content:
        type: string
      title:
        type: string
    type: object
...
  domain.GetSpaceListReq:
    properties:
      app_id:
        type: string
      app_secret:
        type: string
      user_access_token:
        type: string
    type: object
  domain.GetSpaceListResp:
    properties:
      name:
        type: string
      space_id:
        type: string
    type: object
...
  domain.NotnionGetListReq:
    properties:
      cation_title:
        type: string
      integration:
        type: string
    type: object
...
  domain.Page:
    properties:
      content:
        type: string
      id:
        type: string
      parent_id:
        type: string
      title:
        type: string
    type: object
  domain.PageInfo:
    properties:
      id:
        type: string
      title:
        type: string
    type: object
...
  domain.ParseURLItem:
    properties:
      desc:
        type: string
      published:
        type: string
      title:
        type: string
      url:
        type: string
    type: object
  domain.ParseURLReq:
    properties:
      url:
        type: string
    required:
    - url
    type: object
  domain.ParseURLResp:
    properties:
      items:
        items:
          $ref: '#/definitions/domain.ParseURLItem'
        type: array
    type: object
...
  domain.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  domain.ScrapeReq:
    properties:
      kb_id:
        type: string
      url:
        type: string
    required:
    - kb_id
    type: object
  domain.ScrapeResp:
    properties:
      content:
        type: string
      title:
        type: string
    type: object
  domain.SearchDocxReq:
    properties:
      app_id:
        type: string
      app_secret:
        type: string
      user_access_token:
        type: string
    type: object
  domain.SearchDocxResp:
    properties:
      name:
        type: string
      obj_token:
        type: string
      obj_type:
        type: integer
      url:
        type: string
    type: object
  domain.SearchWikiReq:
    properties:
      app_id:
        type: string
      app_secret:
        type: string
      query:
        type: string
      space_id:
        type: string
      user_access_token:
        type: string
    type: object
  domain.SearchWikiResp:
    properties:
      obj_token:
        type: string
      obj_type:
        type: integer
      space_id:
        type: string
      title:
        type: string
      url:
        type: string
    type: object
...
  domain.SiYuanResp:
    properties:
      content:
        type: string
      id:
        type: integer
      title:
        type: string
    type: object
...
  domain.Source:
    properties:
      obj_token:
        type: string
      obj_type:
        type: integer
      url:
        type: string
    type: object
...
  domain.WikiJSResp:
    properties:
      content:
        type: string
      id:
        type: integer
      title:
        type: string
    type: object
  domain.YuqueResp:
    properties:
      content:
        type: string
      title:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/crawler/confluence/analysis_export_file:
    post:
      consumes:
      - application/json
      description: Analyze Confluence Export File
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      - description: kb_id
        in: formData
        name: kb_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.AnalysisConfluenceResp'
                  type: array
              type: object
      summary: AnalysisConfluenceExportFile
      tags:
      - crawler
  /api/v1/crawler/epub/convert:
    post:
      consumes:
      - multipart/form-data
      description: QpubConvert
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      - description: kb_id
        in: formData
        name: kb_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.EpubResp'
              type: object
      summary: QpubConvert
      tags:
      - crawler
  /api/v1/crawler/feishu/get_doc:
    post:
      consumes:
      - application/json
      description: Get Docx in Feishu Spaces
      parameters:
      - description: Get Docx
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.GetDocxReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.GetDocxResp'
                  type: array
              type: object
      summary: FeishuGetDocx
      tags:
      - crawler
  /api/v1/crawler/feishu/list_doc:
    post:
      consumes:
      - application/json
      description: List Docx in Feishu Spaces
      parameters:
      - description: Search Docx
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.SearchDocxReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.SearchDocxResp'
                  type: array
              type: object
      summary: FeishuListDoc
      tags:
      - crawler
  /api/v1/crawler/feishu/list_spaces:
    post:
      consumes:
      - application/json
      description: List All Feishu Spaces
      parameters:
      - description: List Spaces
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.GetSpaceListReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.GetSpaceListResp'
                  type: array
              type: object
      summary: FeishuListSpaces
      tags:
      - crawler
  /api/v1/crawler/feishu/search_wiki:
    post:
      consumes:
      - application/json
      description: Search Wiki in Feishu Spaces
      parameters:
      - description: Search Wiki
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.SearchWikiReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.SearchWikiResp'
                  type: array
              type: object
      summary: FeishuSearchWiki
      tags:
      - crawler
  /api/v1/crawler/notion/get_doc:
    post:
      consumes:
      - application/json
      description: GetDocs
      parameters:
      - description: Get Docs
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.GetDocsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.Page'
                  type: array
              type: object
      summary: GetDocs
      tags:
      - crawler
  /api/v1/crawler/notion/get_list:
    post:
      consumes:
      - application/json
      description: NotionGetList
      parameters:
      - description: Notion Get List
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.NotnionGetListReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.PageInfo'
                  type: array
              type: object
      summary: NotionGetList
      tags:
      - crawler
  /api/v1/crawler/parse_rss:
    post:
      consumes:
      - application/json
      description: Parse RSS
      parameters:
      - description: Parse URL
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.ParseURLReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ParseURLResp'
              type: object
      summary: Parse RSS
      tags:
      - crawler
  /api/v1/crawler/parse_sitemap:
    post:
      consumes:
      - application/json
      description: Parse Sitemap
      parameters:
      - description: Parse URL
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.ParseURLReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ParseURLResp'
              type: object
      summary: Parse Sitemap
      tags:
      - crawler
  /api/v1/crawler/scrape:
    post:
      consumes:
      - application/json
      description: Scrape
      parameters:
      - description: Scrape
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.ScrapeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ScrapeResp'
              type: object
      summary: Scrape
      tags:
      - crawler
  /api/v1/crawler/siyuan/analysis_export_file:
    post:
      consumes:
      - application/json
      description: Analyze SiYuan Export File
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      - description: kb_id
        in: formData
        name: kb_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.SiYuanResp'
                  type: array
              type: object
      summary: AnalysisSiyuanExportFile
      tags:
      - crawler
  /api/v1/crawler/wikijs/analysis_export_file:
    post:
      consumes:
      - multipart/form-data
      description: AnalysisWikijsExportFile
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      - description: kb_id
        in: formData
        name: kb_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.WikiJSResp'
                  type: array
              type: object
      summary: AnalysisWikijsExportFile
      tags:
      - crawler
  /api/v1/crawler/yuque/analysis_export_file:
    post:
      consumes:
      - application/json
      description: Analyze Yuque Export File
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      - description: kb_id
        in: formData
        name: kb_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.YuqueResp'
                  type: array
              type: object
      summary: AnalysisYuqueExportFile
      tags:
      - crawler
...
swagger: "2.0"